Update T function types to rely on i18next types
